Split some functions into reconciler.ts from executor
